requirements.txt
torch>=2.1.0
transformers>=4.35.0
datasets>=2.14.0
tokenizers>=0.14.0
accelerate>=0.24.0
deepspeed>=0.12.0
wandb>=0.16.0
requests>=2.31.0
beautifulsoup4>=4.12.0
scrapy>=2.11.0
selenium>=4.15.0
aiohttp>=3.8.0
pandas>=2.1.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.66.0
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
gradio>=3.50.0
huggingface-hub>=0.17.0
safetensors>=0.4.0
sentencepiece>=0.1.99
gguf>=0.1.0